Fix filter bug add filter logic to PortfolioFilter
