+++ src/components/Portfolio/PortfolioFilter.jsx
@@ -1,8 +1,34 @@
 import React, { useContext, useState } from 'react';
 import GalleryDataContext from '../../reducers/GalleryDataContext';
 import OptionsContext from '../../reducers/OptionsContext';
+import $ from 'jquery';
+import Isotope from 'isotope-layout';
+
 
 const PortfolioFilter = () => {
+  $(document).ready(function () {
+    let grid = document.querySelector('.gridBox');
+    let iso = new Isotope(grid, {
+        itemSelector: '.galleryItem',
+        layoutMode: 'masonry'
+    });
+
+    let filtersElem = document.querySelector('.portfolio-nav');
+    
+    filtersElem.addEventListener('click', function (event) {
+        if (event.target.localName !== "li") {
+            return;
+        }
+
+        let filterValue = event.target.getAttribute('data-filter');
+        
+        iso.arrange({ filter: filterValue });
+    });
+$('.portfolio-nav li').click(function () {
+    $('.portfolio-nav .active').removeClass('active');
+    $(this).addClass('active');
+})
+});
   const { optionsState } = useContext(OptionsContext);
   const { galleryDatadispatch } = useContext(GalleryDataContext);
   const [active, setActive] = useState('All');
@@ -22,7 +48,7 @@ const PortfolioFilter = () => {
           <ul className='portfolio-nav'>
             <li   style={
                 active === 'All' ? { background: optionsState.color } : null
-              } 
+              }
               onClick={() => filtering('All')}
               className='sectionParagraph active'
               data-filter='*'
