+++ src/components/header/HeaderContent.jsx
@@ -62,12 +62,10 @@ const HeaderContent = ({ active, activeHandler }) => {
           </Link>
           <Link
             to="/web-animation-project/portfolio"
-            // reloadDocument
             className={
               activeLink === 3 ? "butList-item active" : "butList-item"
             }
             onClick={() => {
-              window. reload()
               setActiveLink(3);
             }}
           >
