+++ src/style/Portfolio/GalleryItem.scss
@@ -12,6 +12,7 @@
     margin: 0 1.6% 1.6% 1.6% ;
     padding: 0 ;
     background-color: transparent ;
+    transition: all 0.4s;
   
     img {
         width: 100% ;
